+++ src/main/java/com/cinema/ticketbooking/movie/service/IMovieService.java
@@ -8,11 +8,11 @@ public interface IMovieService {
 
     public List<MovieDto> getAllMovies();
 
-    public MovieDto getMovieById(Long id);
+    public MovieDto getMovieById(Integer id);
 
     public MovieDto createMovie(MovieDto movieDto);
 
-    public MovieDto updateMovie(Long id, MovieDto movieDto);
+    public MovieDto updateMovie(Integer id, MovieDto movieDto);
 
 //    public MovieDto getMovieByName(String name);
 //    public boolean deleteMovie(Long id);
