+++ src/main/java/com/cinema/ticketbooking/repository/MovieRepository.java
@@ -3,5 +3,5 @@ package com.cinema.ticketbooking.repository;
 import com.cinema.ticketbooking.entity.Movie;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface MovieRepository extends JpaRepository<Movie, Long> {
+public interface MovieRepository extends JpaRepository<Movie, Integer> {
 }
